Implemented checks in views

+++ api/views.py
@@ -7,6 +7,19 @@ from accounts.serializers import UserSerializer
 invalid_data = {'Error': 'Invalid Data!'}
 
 
+def error_message(key):
+    return Response({"Error": f"Missing or Null value field '{key}' in the request body."})
+
+
+def verify_key(key, request):
+    if key not in request.data:
+        return False
+    elif not request.data.get(key):
+        return False
+    else:
+        return True
+
+
 @api_view(['GET'])
 @permission_classes([AllowAny])
 def api_overview(request):
@@ -30,13 +43,12 @@ def api_overview(request):
 @permission_classes([IsAuthenticated])
 def task_list(request):
     user_id = request.user.id
-    if 'workspace_id' in request.data:
-        workspace_id = request.data.workspace_id
-        task = models.Task.objects.filter(user_id=user_id, workspace_id=workspace_id)
-        serializer = serializers.TaskSerializer(task, many=True)
-        return Response(serializer.data)
-    else:
-        return Response({'Error': 'Need workspace id'})
+    if not verify_key('workspace_id', request):
+        return error_message('workspace_id')
+    workspace_id = request.data.get('workspace_id')
+    task = models.Task.objects.filter(user_id=user_id, workspace_id=workspace_id)
+    serializer = serializers.TaskSerializer(task, many=True)
+    return Response(serializer.data)
 
 
 @api_view(['GET'])
@@ -51,6 +63,17 @@ def task_details(request, pk):
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def task_create(request):
+    request.data['user_id'] = request.user.id
+
+    if not verify_key('workspace_id', request):
+        return error_message('workspace_id')
+
+    workspace_id = request.data.get('workspace_id')
+    workspace = models.Workspace.objects.filter(id=workspace_id)
+
+    if not workspace:
+        return Response({'Error': 'Workspace does not exist'})
+
     serializer = serializers.TaskSerializer(data=request.data)
     if serializer.is_valid(raise_exception=True):
         serializer.save()
@@ -61,8 +84,9 @@ def task_create(request):
 
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
-def task_update(request, pk):
-    task = models.Task.objects.get(id=pk)
+def task_update(request):
+    task_id = request.data.task_id
+    task = models.Task.objects.filter(id=task_id)
     serializer = serializers.TaskSerializer(instance=task, data=request.data)
     if serializer.is_valid(raise_exception=True):
         serializer.save()
@@ -73,9 +97,16 @@ def task_update(request, pk):
 
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated])
-def task_delete(request, pk):
+def task_delete(request):
     user_id = request.user.id
-    task = models.Task.objects.filter(id=pk, user_id=user_id)
+
+    if not verify_key('task_id', request):
+        return error_message('task_id')
+
+    task_id = request.data.get('task_id')
+    task = models.Task.objects.filter(id=task_id, user_id=user_id)
+    if not task:
+        return Response({"Error": "Task does not exist!"})
     task.delete()
     return Response('SUCCESS')
 
@@ -92,9 +123,18 @@ def workspace_list(request):
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def workspace_details(request, pk):
+def workspace_details(request):
     user_id = request.user.id
-    workspace = models.Workspace.objects.filter(id=pk, user_id=user_id)
+
+    if not verify_key('workspace_id', request):
+        return error_message('workspace_id')
+
+    workspace_id = request.data.get('workspace_id')
+    workspace = models.Workspace.objects.filter(id=workspace_id, user_id=user_id)
+
+    if not workspace:
+        return Response({"Error": "No workspace available!"})
+
     serializer = serializers.WorkspaceSerializer(workspace, many=False)
     return Response(serializer.data)
 
@@ -102,6 +142,8 @@ def workspace_details(request, pk):
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def workspace_create(request):
+    user_id = request.user.id
+    request.data['user_id'] = user_id
     serializer = serializers.WorkspaceSerializer(data=request.data)
     if serializer.is_valid(raise_exception=True):
         serializer.save()
@@ -112,8 +154,18 @@ def workspace_create(request):
 
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
-def workspace_update(request, pk):
-    workspace = models.Workspace.objects.get(id=pk)
+def workspace_update(request):
+    user_id = request.user.id
+
+    if not verify_key('workspace_id', request):
+        return error_message('workspace_id')
+
+    workspace_id = request.data.get('workspace_id')
+    workspace = models.Workspace.objects.filter(id=workspace_id, user_id=user_id)
+
+    if not workspace:
+        return Response({'Error': "Workspace does not exist!"})
+
     serializer = serializers.WorkspaceSerializer(instance=workspace, data=request.data)
     if serializer.is_valid(raise_exception=True):
         serializer.save()
@@ -124,8 +176,13 @@ def workspace_update(request, pk):
 
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated])
-def workspace_delete(request, pk):
+def workspace_delete(request):
     user_id = request.user.id
-    workspace = models.Workspace.objects.filter(id=pk, user_id=user_id)
+
+    if not verify_key('workspace_id', request):
+        return error_message('workspace_id')
+
+    workspace_id = request.data.get('workspace_id')
+    workspace = models.Workspace.objects.filter(id=workspace_id, user_id=user_id)
     workspace.delete()
     return Response('SUCCESS')
